WIP rework everything
