+++ front/types/assistant/agent.ts
@@ -1,14 +1,6 @@
-import { ModelId } from "@app/lib/databases";
-import { RetrievalConfigurationType } from "@app/types/assistant/actions/retrieval";
-
 /**
- * Agent Action configuration
+ * Agent Action
  */
-
-// New AgentActionConfigurationType checklist:
-// - Add the type to the union type below
-// - Add model rendering support in `renderConversationForModel`
-export type AgentActionConfigurationType = RetrievalConfigurationType;
 
 // Each AgentActionConfigurationType is capable of generating this type at runtime to specify which
 // inputs should be generated by the model. As an example, to run the retrieval action for which the
@@ -32,7 +24,6 @@ export type AgentActionConfigurationType = RetrievalConfigurationType;
 // ```
 
 export type AgentActionSpecification = {
-  id: ModelId;
   name: string;
   description: string;
   inputs: {
@@ -40,41 +31,4 @@ export type AgentActionSpecification = {
     description: string;
     type: "string" | "number" | "boolean";
   }[];
-};
-
-/**
- * Agent Message configuration
- */
-
-export type AgentGenerationConfigurationType = {
-  id: ModelId;
-  prompt: string;
-  model: {
-    providerId: string;
-    modelId: string;
-  };
-};
-
-/**
- * Agent configuration
- */
-
-export type AgentConfigurationStatus = "active" | "archived";
-export type AgentConfigurationScope = "global" | "workspace";
-
-export type AgentConfigurationType = {
-  id: ModelId;
-  sId: string;
-  status: AgentConfigurationStatus;
-  name: string;
-  pictureUrl: string | null;
-};
-
-export type AgentFullConfigurationType = {
-  agent: AgentConfigurationType;
-  // If undefined, no action performed, otherwise the action is
-  // performed (potentially NoOp eg autoSkip above).
-  action: AgentActionConfigurationType | null;
-  // If undefined, no text generation.
-  generation: AgentGenerationConfigurationType | null;
 };
